+++ AzeriteUI/modules/unitframe_arenaframes.lua
@@ -5,10 +5,9 @@ if (not AzeriteUI) then
 	return 
 end
 
-
 local UnitFrameArena = AzeriteUI:NewModule("UnitFrameArena", "LibDB", "LibEvent", "LibUnitFrame", "LibStatusBar")
 local Colors = CogWheel("LibDB"):GetDatabase("AzeriteUI: Colors")
-local WhiteList = CogWheel("LibDB"):GetDatabase("AzeriteUI: Auras").WhiteList
+local Auras = CogWheel("LibDB"):GetDatabase("AzeriteUI: Auras")
 
 -- Lua API
 local _G = _G
@@ -16,6 +15,26 @@ local unpack = unpack
 
 -- WoW Strings
 local DEAD = _G.DEAD
+
+-- Cast Bar Map
+-- (Texture Size 128x32, Growth: RIGHT)
+local map = {
+	cast = {
+		top = {
+			{ keyPercent =   0/128, offset = -16/32 }, 
+			{ keyPercent =  10/128, offset =   0/32 }, 
+			{ keyPercent = 119/128, offset =   0/32 }, 
+			{ keyPercent = 128/128, offset = -16/32 }
+		},
+		bottom = {
+			{ keyPercent =   0/128, offset = -16/32 }, 
+			{ keyPercent =  10/128, offset =   0/32 }, 
+			{ keyPercent = 119/128, offset =   0/32 }, 
+			{ keyPercent = 128/128, offset = -16/32 }
+		}
+	}
+
+}
 
 
 -- Utility Functions
@@ -47,55 +66,6 @@ local OverrideHealthValue = function(element, unit, min, max, disconnected, dead
 	else 
 		return OverrideValue(element, unit, min, max, disconnected, dead, tapped)
 	end 
-end 
-
-
-local BuffFilter = function(element, button, unit, isOwnedByPlayer, name, icon, count, debuffType, duration, expirationTime, unitCaster, isStealable, nameplateShowPersonal, spellId, canApplyAura, isBossDebuff, isCastByPlayer, nameplateShowAll, timeMod, value1, value2, value3)
-
-	-- ALways whitelisted auras, boss debuffs and stealable for mages
-	if WhiteList[spellId] or isBossDebuff or (PlayerClass == "MAGE" and isStealable) then 
-		return true 
-	end 
-
-	-- Try to hide non-player auras outdoors
-	if (not isOwnedByPlayer) and (not IsInInstance()) then 
-		return 
-	end 
-
-	-- Hide static and very long ones
-	if (not duration) or (duration > 60) then 
-		return 
-	end 
-
-	-- show our own short ones
-	if (isOwnedByPlayer and duration and (duration > 0) and (duration < 60)) then 
-		return true
-	end 
-	
-	
-end 
-
-local DebuffFilter = function(element, button, unit, isOwnedByPlayer, name, icon, count, debuffType, duration, expirationTime, unitCaster, isStealable, nameplateShowPersonal, spellId, canApplyAura, isBossDebuff, isCastByPlayer, nameplateShowAll, timeMod, value1, value2, value3)
-
-	if WhiteList[spellId] or isBossDebuff then 
-		return true 
-	end 
-
-	-- Try to hide non-player auras outdoors
-	if (not isOwnedByPlayer) and (not IsInInstance()) then 
-		return 
-	end 
-
-	-- Hide static and very long ones
-	if (not duration) or (duration > 60) then 
-		return 
-	end 
-
-	-- show our own short ones
-	if (isOwnedByPlayer and duration and (duration > 0) and (duration < 60)) then 
-		return true
-	end 
-
 end 
 
 local PostCreateAuraButton = function(element, button)
@@ -175,7 +145,7 @@ local Style = function(self, unit, id, ...)
 		id = counter 
 	end 
 
-	local width, height = 96, 36
+	local width, height = 136, 47 -- 96, 36
 	local spacing = 30 + 14 + 6
 
 	self:SetSize(width, height)
@@ -208,13 +178,14 @@ local Style = function(self, unit, id, ...)
 	-----------------------------------------------------------	
 
 	local health = content:CreateStatusBar()
-	health:SetSize(75, 13)
-	health:Place("BOTTOM", 0, 0)
+	health:SetSize(111,14)
+	health:Place("CENTER", 0, 0)
 	health:SetOrientation("LEFT") -- set the bar to grow towards the right.
 	health:SetSmoothingMode("bezier-fast-in-slow-out") -- set the smoothing mode.
 	health:SetSmoothingFrequency(.5) -- set the duration of the smoothing.
 	health:SetStatusBarTexture(getPath("cast_bar"))
-	health.colorTapped = false -- color tap denied units 
+	health:SetSparkMap(map.cast) -- set the map the spark follows along the bar.
+	health.colorTapped = true -- color tap denied units 
 	health.colorDisconnected = true -- color disconnected units
 	health.colorClass = true -- color players by class 
 	health.colorReaction = true -- color NPCs by their reaction standing with us
@@ -224,8 +195,8 @@ local Style = function(self, unit, id, ...)
 
 	local healthBg = health:CreateTexture()
 	healthBg:SetDrawLayer("BACKGROUND", -1)
-	healthBg:SetSize(130, 84)
-	healthBg:SetPoint("CENTER", 0, -2)
+	healthBg:SetSize(193,93)
+	healthBg:SetPoint("CENTER", 1, -2)
 	healthBg:SetTexture(getPath("cast_back"))
 	healthBg:SetVertexColor(Colors.ui.stone[1], Colors.ui.stone[2], Colors.ui.stone[3])
 	self.Health.Bg = healthBg
@@ -235,10 +206,11 @@ local Style = function(self, unit, id, ...)
 	-----------------------------------------------------------	
 
 	local absorb = content:CreateStatusBar()
-	absorb:SetSize(75, 13)
+	absorb:SetSize(111,14)
 	absorb:SetStatusBarTexture(getPath("cast_bar"))
+	absorb:SetSparkMap(map.cast) -- set the map the spark follows along the bar.
 	absorb:SetFrameLevel(health:GetFrameLevel() + 2)
-	absorb:Place("BOTTOM", 0, 0)
+	absorb:Place("CENTER", 0, 0)
 	absorb:SetOrientation("RIGHT") -- grow the bar towards the left (grows from the end of the health)
 	absorb:SetStatusBarColor(1, 1, 1, .25) -- make the bar fairly transparent, it's just an overlay after all. 
 	self.Absorb = absorb
@@ -247,14 +219,16 @@ local Style = function(self, unit, id, ...)
 	-- Cast Bar
 	-----------------------------------------------------------
 	local cast = content:CreateStatusBar()
-	cast:SetSize(75, 13)
+	cast:SetSize(111,14)
 	cast:SetStatusBarTexture(getPath("cast_bar"))
+	cast:SetSparkMap(map.cast) -- set the map the spark follows along the bar.
 	cast:SetFrameLevel(health:GetFrameLevel() + 1)
-	cast:Place("BOTTOM", 0, 0)
+	cast:Place("CENTER", 0, 0)
 	cast:SetOrientation("LEFT") 
 	cast:SetStatusBarColor(1, 1, 1, .15) 
 	cast:DisableSmoothing(true) 
 	self.Cast = cast
+
 
 
 	-- Auras
@@ -262,14 +236,17 @@ local Style = function(self, unit, id, ...)
 
 	local auras = content:CreateFrame("Frame")
 	auras:Place("RIGHT", health, "LEFT", -26, -1)
-	auras:SetSize(36*6 + 8*5, 36) -- auras will be aligned in the available space, this size gives us 7x1 auras
+	auras:SetSize(36*6 + 8*5, 36) -- auras will be aligned in the available space
 
 	auras.auraSize = 34 -- too much?
 	auras.spacingH = 4 -- horizontal/column spacing between buttons
 	auras.spacingV = 4 -- vertical/row spacing between aura buttons
 	auras.growthX = "LEFT" -- auras grow to the left
 	auras.growthY = "DOWN" -- rows grow downwards (we just have a single row, though)
-	auras.maxButtons = nil -- when set will limit the number of buttons regardless of space available
+	auras.maxVisible = 6 -- when set will limit the number of buttons regardless of space available
+	auras.maxBuffs = 3 -- maximum number of visible buffs
+	auras.maxDebuffs = nil -- maximum number of visible debuffs
+	auras.debuffsFirst = false -- show debuffs before buffs
 	auras.showCooldownSpiral = false -- don't show the spiral as a timer
 	auras.showCooldownTime = true -- show timer numbers
 
@@ -280,9 +257,9 @@ local Style = function(self, unit, id, ...)
 
 	-- Filter methods
 	auras.AuraFilter = nil -- general aura filter function, called when the below aren't there
-	auras.BuffFilter = BuffFilter -- buff specific filter function
-	auras.DebuffFilter = DebuffFilter -- debuff specific filter function
-
+	auras.BuffFilter = Auras.BuffFilter -- buff specific filter function
+	auras.DebuffFilter = Auras.DebuffFilter -- debuff specific filter function
+		
 	-- Aura tooltip position
 	auras.tooltipDefaultPosition = nil 
 	auras.tooltipPoint = "TOPRIGHT"
@@ -301,7 +278,7 @@ local Style = function(self, unit, id, ...)
 
 	-- Unit Name
 	local name = overlay:CreateFontString()
-	name:SetPoint("BOTTOMRIGHT", health, "TOPRIGHT", 0, 16)
+	name:SetPoint("BOTTOMRIGHT", health, "TOPRIGHT", -10, 16)
 	name:SetDrawLayer("OVERLAY")
 	name:SetJustifyH("CENTER")
 	name:SetJustifyV("TOP")
@@ -312,13 +289,13 @@ local Style = function(self, unit, id, ...)
 	self.Name = name
 
 	local healthVal = overlay:CreateFontString()
-	healthVal:SetPoint("CENTER", health, "CENTER", 0, 0)
+	healthVal:SetPoint("RIGHT", health, "RIGHT", -10, 0)
 	healthVal:SetDrawLayer("OVERLAY")
 	healthVal:SetJustifyH("CENTER")
 	healthVal:SetJustifyV("MIDDLE")
-	healthVal:SetFontObject(AzeriteFont11_Outline)
-	healthVal:SetShadowOffset(-.85, -.85)
-	healthVal:SetShadowColor(0, 0, 0, .75)
+	healthVal:SetFontObject(AzeriteFont14_Outline)
+	healthVal:SetShadowOffset(0, 0)
+	healthVal:SetShadowColor(0, 0, 0, 0)
 	healthVal:SetTextColor(240/255, 240/255, 240/255, .5)
 	self.Health.Value = healthVal
 	self.Health.OverrideValue = OverrideHealthValue